+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Version _next_
+## Version 1.3.0 - 2020-07-12
 
 - Add support for generic JSON data for properties.
 - Add support for `writeOnly`.